backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
PyYAML>=6.0
httpx>=0.26
